backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
pydantic
httpx
google-generativeai
